requirements.txt
torch
transformers
bs4
lxml
pandas
python-docx
